requirements.txt
streamlit==1.36.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
langchain-experimental==0.0.62
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langchainhub==0.1.20
python-dotenv==1.0.1
pandas==1.5.3
matplotlib
numpy